% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{tidymod_lm_matrix}
\alias{tidymod_lm_matrix}
\title{OLS estimation - Matrix}
\usage{
tidymod_lm_matrix(X, Y, ...)
}
\arguments{
\item{X}{is the covariate matrix, which contains all explanatory variables. Has to be a matrix or be convertible
with as.matrix()}

\item{Y}{is the explained variable. Has to be numeric or be convertible with as.numeric()}
}
\value{
A list is returned containing the following elements coefficients, standard errors,
degrees of freedom, variance-covariance matrix, fitted values, residuals, data used in the regression,
the call, the intercept and the formula
}
\description{
The function conduct the very common OLS estimation, which minimizes the sum of squared residuals.
}
\examples{
tidymod_lm_matrix(mtcars[,2:3], mtcars[,1])
}
