\name{tidymodelR-package}
\alias{tidymodelR-package}
\alias{tidymodelR}
\docType{package}
\title{
\packageTitle{tidymodelR}
}
\description{
\packageDescription{tidymodelR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tidymodelR}
\packageIndices{tidymodelR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tidymodelR}

Maintainer: \packageMaintainer{tidymodelR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
