% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodelR.R
\docType{package}
\name{tidymodelR-package}
\alias{tidymodelR}
\alias{tidymodelR-package}
\title{tidymodelR: Statistical modelling and testing}
\description{
A implementation of modelling and testing using piping.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elben10.github.io/tidymodelR}
  \item Report bugs at \url{https://github.com/elben10/tidymodelR/issues}
}

}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com}

}
