% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodelR.R
\docType{package}
\name{tidymodelR-package}
\alias{tidymodelR}
\alias{tidymodelR-package}
\title{tidymodelR: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com}

}
